LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/cors-crossorigin-requests-expected.txt

CONSOLE MESSAGE: SyntaxError: JSON Parse error: Expected '}'
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
json-module-crossorigin



PASS Imported JSON module, cross-origin with CORS
PASS Imported JSON module, cross-origin, missing CORS ACAO header
PASS Imported JSON module with parse error, cross-origin, with CORS
